feat: cache handler revalidate tag functionality

+++ packages/redis-cache/src/index.ts
@@ -1,3 +1,4 @@
+import { NEXT_CACHE_IMPLICIT_TAG_ID, NEXT_CACHE_TAGS_HEADER } from 'next/dist/lib/constants'
 import { createClient, RedisClientType, RedisClientOptions } from 'redis'
 import type { CacheEntry, CacheStrategy } from '@dbbs/next-cache-handler-common'
 
@@ -9,25 +10,52 @@ export class RedisCache implements CacheStrategy {
     this.client.connect()
   }
 
-  async get(_pageKey: string, cacheKey: string): Promise<CacheEntry | null> {
-    const pageData = await this.client.get(cacheKey)
+  async get(pageKey: string, cacheKey: string): Promise<CacheEntry | null> {
+    const pageData = await this.client.get(`${pageKey}//${cacheKey}`)
     if (!pageData) return null
     return JSON.parse(pageData)
   }
 
-  async set(_pageKey: string, cacheKey: string, data: CacheEntry): Promise<void> {
-    await this.client.set(cacheKey, JSON.stringify(data))
+  async set(pageKey: string, cacheKey: string, data: CacheEntry): Promise<void> {
+    await this.client.set(`${pageKey}//${cacheKey}`, JSON.stringify(data), {
+      ...(data.revalidate ? { EX: Number(data.revalidate) } : undefined)
+    })
   }
 
-  async delete(_pageKey: string, cacheKey: string): Promise<void> {
-    await this.client.del(cacheKey)
+  async revalidateTag(tag: string): Promise<void> {
+    let cursor = 0
+    do {
+      const { cursor: currentCursor, keys } = await this.client.scan(0, { MATCH: '*', COUNT: 100 })
+      cursor = currentCursor
+      for (const cacheKey of keys) {
+        const data = await this.client.get(cacheKey)
+        if (!data) continue
+
+        const pageData: CacheEntry | null = JSON.parse(data)
+        if (
+          pageData?.tags?.includes(tag) ||
+          (pageData?.value?.kind === 'PAGE' &&
+            pageData.value?.headers?.[NEXT_CACHE_TAGS_HEADER]?.toString()?.split(',').includes(tag))
+        ) {
+          await this.client.del(cacheKey)
+        }
+      }
+    } while (cursor)
+    return
+  }
+
+  async delete(pageKey: string, cacheKey: string): Promise<void> {
+    await this.client.del(`${pageKey}//${cacheKey}`)
   }
 
   async deleteAllByKeyMatch(key: string): Promise<void> {
-    const allKeys = this.client.scanIterator({ MATCH: key })
-
-    for await (const cacheKey of allKeys) {
-      this.client.del(cacheKey)
-    }
+    let cursor = 0
+    do {
+      const result = await this.client.scan(0, { MATCH: `${key}//*`, COUNT: 100 })
+      cursor = result.cursor
+      for await (const cacheKey of result.keys) {
+        await this.client.del(cacheKey)
+      }
+    } while (cursor != 0)
   }
 }
